Extract Prompt Studio node UI hook
