+++ docs/development/issue-14-frontend-js-maintenance.md
@@ -229,6 +229,7 @@ web/js/
     advanced_layout_controller.js
     studio_textareas.js
     studio_resizable_input.js
+    studio_node_ui.js
     studio_values.js
     wildcard_values.js
     textarea.js
@@ -767,6 +768,7 @@ Use this in Issue #14 or follow-up PR tracking comments.
 [ ] advanced_layout_controller.js extracted
 [ ] studio_textareas.js extracted
 [ ] studio_resizable_input.js extracted
+[ ] studio_node_ui.js extracted
 [ ] studio_values.js extracted
 [ ] wildcard_values.js extracted
 [ ] textarea.js extracted
